requirements.txt
# pip install -r requirements.txt

streamlit
# sqlite3
openai
# os
setuptools #para dotenv
wheel #para dotenv
python-dotenv